requirements.txt
aiogram==2.25.1
aiohttp==3.8.5
aiosignal==1.3.1
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
Babel==2.9.1
billiard==3.6.4.0
celery==5.2.7
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==4.2.3
django-admin-interface==0.26.0
django-colorfield==0.9.0
django-js-asset==2.0.0
django-json-widget==1.1.1
django-modeltranslation==0.18.10
django-tinymce==3.6.1
djangorestframework==3.14.0
dnspython==2.3.0
drf-standardized-errors==0.12.5
frozenlist==1.3.3
future==0.18.3
gunicorn==20.1.0
idna==3.4
Jinja2==3.1.2
kombu==5.2.4
magic-filter==1.0.9
MarkupSafe==2.1.3
multidict==6.0.4
Pillow==9.5.0
prompt-toolkit==3.0.38
psycopg==3.1.9
pymongo==4.4.0
python-dotenv==1.0.0
python-slugify==8.0.1
pytz==2023.3
redis==4.5.5
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
typing_extensions==4.6.3
urllib3==2.0.2
vine==5.0.0
wcwidth==0.2.6
yarl==1.9.2
